+++ packages/client/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.4.0
 * Added more example types (`bool`, `exists`, `date`, `geo`, `dateHistogram`)
 * Add support for default properties on node types
+* Add default contexts for `facet` and `results`
 
 # 2.3.0
 * Added lens